requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.27.2
boto3==1.22.4
channels==3.0.4
channels_redis==3.0.0
scikit-learn==1.0.2
pandas==1.4.0
matplotlib==3.5.1
joblib==1.1.0
celery==5.2.3  # Optional: for background tasks
psycopg2-binary==2.9.3  # For PostgreSQL
django-cors-headers==3.14.0
python-magic==0.4.27  # For file type detection
pyotp==2.9.0  # For TOTP 2FA
cryptography==41.0.0  # For data encryption
